修改bug、修改设计图
1、修改未登录收藏响应异常问题
2、增加专家名榜、精图排行、专题精选设计图

+++ Application/Home/Controller/CollectController.class.php
@@ -53,28 +53,33 @@ class CollectController extends LayoutController {
         $this->display();
 	}
 	
+	//收藏文章
 	public function collect() {
-		$Collect = M("Collect");
-		$condition['user_id'] = array('eq',session('userId'));
-		$condition['article_id'] = array('eq',$_POST['articleId']);
-		$collectData = $Collect->where($condition)->find();
-		if($collectData) {
-			$result = $Collect->where($condition)->delete();
-			if($result){
-				$this->ajaxReturn('no');
+		if(cookie('PHPSESSID') && session('id') && cookie('PHPSESSID') == session('id')) {
+			$Collect = M("Collect");
+			$condition['user_id'] = array('eq',session('userId'));
+			$condition['article_id'] = array('eq',$_POST['articleId']);
+			$collectData = $Collect->where($condition)->find();
+			if($collectData) {
+				$result = $Collect->where($condition)->delete();
+				if($result){
+					$this->ajaxReturn('no');
+				} else {
+					$this->ajaxReturn('error');
+				}
 			} else {
-				$this->ajaxReturn('error');
+				$data['user_id'] = session('userId');
+				$data['article_id'] = $_POST['articleId'];
+				$data['createtime'] = date('Y-m-d H:i:s');;
+				$result = $Collect->add($data);
+				if($result){
+					$this->ajaxReturn('yes');
+				} else {
+					$this->ajaxReturn('error');
+				}
 			}
 		} else {
-			$data['user_id'] = session('userId');
-			$data['article_id'] = $_POST['articleId'];
-			$data['createtime'] = date('Y-m-d H:i:s');;
-			$result = $Collect->add($data);
-			if($result){
-				$this->ajaxReturn('yes');
-			} else {
-				$this->ajaxReturn('error');
-			}
+			$this->ajaxReturn('permission');
 		}	
 	}
 }
